backend/services/agent_controller/requirements.txt
celery==5.5.3
fastapi==0.111.0
jsonschema==4.23.0
openai==1.93.0
pydantic==2.11.5
pymysql==1.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==6.2.0
requests==2.31.0
sqlalchemy==2.0.29
uvicorn==0.28.1
watchdog===6.0.0